Move widget modules to separate folder
The multipurpose basic widget modules are moved into a separate folder/package, excluding `widgets.py`.
