+++ app/widgets/__init__.py
@@ -1,4 +1,2 @@
 from . import game
-from . import button
-from . import fps_counter
-from . import slider
+from . import basic
