+++ app/widgets/basic/__init__.py
@@ -1,0 +1,4 @@
+from . import button
+from . import slider
+from . import panel
+from . import fps_counter
