+++ app/widgets/basic/fps_counter.py
