+++ app/widgets/game/bet_prompt.py
@@ -4,8 +4,8 @@ from app.animations.anim_group import AnimGroup
 from app.animations.interpolations import ease_in, ease_out
 from app.animations.move import MoveAnimation
 from app.widgets.game.action_buttons import COLORS, SideTextedButton
-from app.widgets.button import Button
-from app.widgets.slider import Slider
+from app.widgets.basic.button import Button
+from app.widgets.basic.slider import Slider
 from app.shared import *
 from app.widgets.widget import Widget
